# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Student Admission & Entry Test Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /violations:
    get:
      operationId: listViolations
      tags: [violations]
      summary: List anti-cheating violations
      parameters:
        - in: query
          name: sessionId
          schema:
            type: integer
        - in: query
          name: studentId
          schema:
            type: integer
        - in: query
          name: type
          schema:
            type: string
        - in: query
          name: page
          schema:
            type: integer
        - in: query
          name: limit
          schema:
            type: integer
      responses:
        "200":
          description: Violations list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ViolationListResponse"
    post:
      operationId: reportViolation
      tags: [violations]
      summary: Report a cheating violation during test
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ViolationInput"
      responses:
        "201":
          description: Violation logged
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Violation"

  # Dashboard
...
components:
  schemas:
...
    Violation:
      type: object
      required: [id, sessionId, studentId, type, severity, createdAt]
      properties:
        id:
          type: integer
        sessionId:
          type: integer
        studentId:
          type: integer
        studentName:
          type: ["string", "null"]
        rollNumber:
          type: ["string", "null"]
        type:
          type: string
        severity:
          type: string
          enum: [low, medium, high]
        capturedImageUrl:
          type: ["string", "null"]
        faceMatchScore:
          type: ["number", "null"]
        description:
          type: ["string", "null"]
        createdAt:
          type: string
    ViolationInput:
      type: object
      required: [sessionId, studentId, type, severity]
      properties:
        sessionId:
          type: integer
        studentId:
          type: integer
        type:
          type: string
        severity:
          type: string
          enum: [low, medium, high]
        capturedImageBase64:
          type: string
        faceMatchScore:
          type: number
        description:
          type: string
    ViolationListResponse:
      type: object
      required: [data, total, page, limit]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Violation"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer

    # Dashboard
...