# lib/api-zod/src/generated/types/violationInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Student Admission & Entry Test Management System API
 * OpenAPI spec version: 0.1.0
 */
import type { ViolationInputSeverity } from './violationInputSeverity';

export interface ViolationInput {
  sessionId: number;
  studentId: number;
  type: string;
  severity: ViolationInputSeverity;
  capturedImageBase64?: string;
  faceMatchScore?: number;
  description?: string;
}